Category id 와 parentNo 을 Integer Type -> Long Type 으로 변

+++ src/main/java/com/mall/commerce/entity/Category.java
@@ -13,13 +13,13 @@ public class Category {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name= "category_no")
-    private Integer id;
+    private Long id;
 
     @Column(name = "category_name")
     private String categoryName;
 
     @Column(name = "parent_no")
-    private Integer parentNo;
+    private Long parentNo;
 
     @Column(name = "depth")
     private String depth;
